chore: add initial build workflow
