+++ go.mod
@@ -6,27 +6,30 @@ require (
 	github.com/AlecAivazis/survey/v2 v2.2.12
 	github.com/Jeffail/gabs/v2 v2.6.1
 	github.com/MakeNowJust/heredoc v1.0.0
+	github.com/briandowns/spinner v1.18.1
 	github.com/cli/browser v1.1.0
+	github.com/containerd/containerd v1.6.3 // indirect
+	github.com/containerd/continuity v0.3.0 // indirect
 	github.com/deepsourcelabs/graphql v0.2.2
-	github.com/fatih/color v1.12.0
+	github.com/docker/distribution v2.8.1+incompatible // indirect
+	github.com/docker/docker v20.10.14+incompatible
+	github.com/fatih/color v1.13.0
 	github.com/getsentry/sentry-go v0.6.0
-	github.com/go-playground/validator/v10 v10.10.1 // indirect
-	github.com/google/go-cmp v0.5.5 // indirect
-	github.com/kr/pretty v0.3.0 // indirect
+	github.com/go-playground/validator/v10 v10.10.1
+	github.com/gorilla/mux v1.8.0 // indirect
+	github.com/magiconair/properties v1.8.6 // indirect
 	github.com/matryer/is v1.4.0 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/mitchellh/mapstructure v1.4.1 // indirect
+	github.com/moby/sys/mount v0.3.2 // indirect
+	github.com/moby/term v0.0.0-20210619224110-3f7ff695adc6 // indirect
 	github.com/owenrumney/go-sarif/v2 v2.1.0
-	github.com/pelletier/go-toml v1.9.2
-	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pelletier/go-toml v1.9.4
 	github.com/pterm/pterm v0.12.23
-	github.com/rogpeppe/go-internal v1.8.0 // indirect
-	github.com/spf13/cobra v1.1.3
-	github.com/spf13/viper v1.7.1
+	github.com/spf13/afero v1.8.2 // indirect
+	github.com/spf13/cobra v1.4.0
+	github.com/spf13/viper v1.10.1
 	github.com/stretchr/testify v1.7.0
+	github.com/xeipuuv/gojsonpointer v0.0.0-20190905194746-02993c407bfb // indirect
 	github.com/xeipuuv/gojsonschema v1.2.0
-	golang.org/x/sys v0.0.0-20211205182925-97ca703d548d // indirect
-	golang.org/x/text v0.3.7 // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+	gopkg.in/ini.v1 v1.66.4 // indirect
+	gotest.tools/v3 v3.2.0 // indirect
 )
